+++ xmake/actions/require/impl/actions/download.lua
@@ -94,12 +94,13 @@ function _download(package, url, sourcedir, url_alias, url_excludes)
     --
     -- we need not sourcehash and skip checksum to try download it directly if no version list in package()
     -- @see https://github.com/xmake-io/xmake/issues/930
+    -- https://github.com/xmake-io/xmake/issues/1009
     --
     local sourcehash = package:sourcehash(url_alias)
     if sourcehash then
         sourcehash = sourcehash:lower()
     end
-    assert(not package:get("versions") or sourcehash, "cannot get source hash of %s in package(%s)", url, package:name())
+    assert(not package:verify() or not package:get("versions") or sourcehash, "cannot get source hash of %s in package(%s)", url, package:name())
 
     -- the package file have been downloaded?
     local cached = true
@@ -158,7 +159,7 @@ function _urls(package)
     for _, url in ipairs(package:urls()) do
         if git.checkurl(url) then
             table.insert(urls[1], url)
-        elseif not package:get("versions") or package:sourcehash(package:url_alias(url)) then
+        elseif not package:verify() or not package:get("versions") or package:sourcehash(package:url_alias(url)) then
             table.insert(urls[2], url)
         end
     end