+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add new [xrepo](https://github.com/xmake-io/xrepo) command to manage C/C++ packages
 * Support for installing packages of cross-compilation
 * Add musl.cc toolchains
+* [#1009](https://github.com/xmake-io/xmake/issues/1009): Support select and install any version package, e.g. `add_requires("libcurl 7.73.0", {verify = false})`
 
 ### Change
 
@@ -860,6 +861,7 @@
 * 添加新的 [xrepo](https://github.com/xmake-io/xrepo) 命令去管理安装 C/C++ 包
 * 支持安装交叉编译的依赖包
 * 新增musl.cc上的工具链支持
+* [#1009](https://github.com/xmake-io/xmake/issues/1009): 支持忽略校验去安装任意版本的包，`add_requires("libcurl 7.73.0", {verify = false})`
 
 ### 改进